# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "캠페인 추천 시스템 API",
        "title": "ReviewMaps API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
...
        "/keyword-alerts/alerts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "List user alerts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.AlertListResponse"
                        }
                    }
                }
            }
        },
        "/keyword-alerts/alerts/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Mark alerts as read",
                "parameters": [
                    {
                        "description": "Alert IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.MarkReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/keyword-alerts/alerts/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Delete alert",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/keyword-alerts/fcm/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Register FCM token",
                "parameters": [
                    {
                        "description": "FCM token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.FCMRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FCMDevice"
                        }
                    }
                }
            }
        },
        "/keyword-alerts/fcm/unregister": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Unregister FCM token",
                "parameters": [
                    {
                        "description": "FCM token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.FCMUnregisterRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/keyword-alerts/keywords": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "List user keywords",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Keyword"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Create keyword",
                "parameters": [
                    {
                        "description": "Keyword data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateKeywordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Keyword"
                        }
                    }
                }
            }
        },
        "/keyword-alerts/keywords/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Delete keyword",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Keyword ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/keyword-alerts/keywords/{id}/toggle": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keyword-alerts"
                ],
                "summary": "Toggle keyword active status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Keyword ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Keyword"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Campaign": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "apply_deadline": {
                    "type": "string"
                },
                "apply_from": {
                    "type": "string"
                },
                "campaign_channel": {
                    "type": "string"
                },
                "campaign_type": {
                    "type": "string"
                },
                "category": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ]
                },
                "category_id": {
                    "type": "integer"
                },
                "company": {
                    "type": "string"
                },
                "company_link": {
                    "type": "string"
                },
                "content_link": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "offer": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "promotion_level": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "review_deadline": {
                    "type": "string"
                },
                "search_text": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "campaigns": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Campaign"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.FCMDevice": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Keyword": {
            "type": "object",
            "properties": {
                "alerts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.KeywordAlert"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "keyword": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.KeywordAlert": {
            "type": "object",
            "properties": {
                "campaign": {
                    "$ref": "#/definitions/models.Campaign"
                },
                "campaign_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "is_sent": {
                    "type": "boolean"
                },
                "keyword": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Keyword"
                        }
                    ]
                },
                "keyword_id": {
                    "type": "integer"
                }
            }
        },
        "models.SocialAccount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "date_joined": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fcm_devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FCMDevice"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_staff": {
                    "type": "boolean"
                },
                "is_superuser": {
                    "type": "boolean"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Keyword"
                    }
                },
                "last_login": {
                    "type": "string"
                },
                "login_method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_image": {
                    "type": "string"
                },
                "social_accounts": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SocialAccount"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "services.AlertListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.KeywordAlert"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "services.CreateKeywordRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                }
            }
        },
        "services.FCMRegisterRequest": {
            "type": "object",
            "properties": {
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "services.FCMUnregisterRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "services.MarkReadRequest": {
            "type": "object",
            "properties": {
                "alert_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}